+++ packages/user-management-widget/src/lib/widget/drivers/BaseDriver.ts
@@ -1,22 +1,6 @@
 import { Logger } from '../../mixins/loggerMixin/types';
-
-type Empty = null | undefined;
-type RefOrRefFn = Element | (() => HTMLElement | Empty) | Empty;
-
-const waitForElement = async (ele: RefOrRefFn, timeout: number) =>
-  new Promise<Element | null>((resolve) => {
-    const interval = setInterval(() => {
-      const element = typeof ele === 'function' ? ele() : ele;
-      if (element) {
-        clearInterval(interval);
-        resolve(element);
-      }
-    }, 100);
-    setTimeout(() => {
-      clearInterval(interval);
-      resolve(null);
-    }, timeout);
-  });
+import { waitForElement } from './helpers';
+import { RefOrRefFn } from './types';
 
 export class BaseDriver {
   #ele: RefOrRefFn;
