+++ packages/user-management-widget/src/lib/widget/drivers/types.ts
@@ -1,0 +1,3 @@
+type Empty = null | undefined;
+
+export type RefOrRefFn = Element | (() => HTMLElement | Empty) | Empty;
